Add some delete dialog

+++ lib/ui/admin_screen/admin_jemaatdata_screen.dart
@@ -148,7 +148,36 @@ class _AdminJemaatDataScreenState extends State<AdminJemaatDataScreen> {
                       color: Colors.red,
                       icon: Icons.delete,
                       onTap: () {
-                        DatabaseJemaatServices.deleteData(e.id);
+                        showDialog(
+                            context: context,
+                            builder: (context) {
+                              return AlertDialog(
+                                title: Text(
+                                    AppLocalizations.of(context)!.alertWarning),
+                                content: Text(AppLocalizations.of(context)!
+                                    .alertWarningMessageDelete),
+                                actions: [
+                                  TextButton(
+                                    child: Text(
+                                      AppLocalizations.of(context)!.noButton,
+                                      style: TextStyle(color: secondaryColor),
+                                    ),
+                                    onPressed: () {
+                                      Navigator.pop(context);
+                                    },
+                                  ),
+                                  TextButton(
+                                    child: Text(
+                                      AppLocalizations.of(context)!.yesButton,
+                                      style: TextStyle(color: secondaryColor),
+                                    ),
+                                    onPressed: () {
+                                      DatabaseJemaatServices.deleteData(e.id);
+                                    },
+                                  )
+                                ],
+                              );
+                            });
                       },
                     )
                   ],
